GRC/Tools/DLP/requirements.txt
pyyaml
yagmail
